Renovation: preact-wrapper folder refactoring (#13804)
* Renovation: preact-wrapper folder refactoring

* ignore ts errors

* fix strange generator error
