--- js/renovation/preact-wrapper/button.d.ts
@@ -1,13 +1,0 @@
-/* eslint-disable no-underscore-dangle */
-import Component from './component';
-
-export default class Button extends Component {
-  getProps(): any;
-
-  _init(): any;
-
-  _getSubmitAction(): any;
-
-  _findGroup(): any;
-}
-/* eslint-enable no-underscore-dangle */
