--- js/renovation/preact-wrapper/tooltip-item-content.d.ts
@@ -1,7 +1,0 @@
-/* eslint-disable no-underscore-dangle */
-import Component from './component';
-
-export default class TooltipItemContent extends Component {
-    _setOptionsByReference(): void;
-}
-/* eslint-enable no-underscore-dangle */
