--- js/renovation/preact-wrapper/tooltip-item-content.js
@@ -1,12 +1,0 @@
-import Component from './component';
-import { extend } from '../../core/utils/extend';
-
-export default class TooltipItemContent extends Component {
-    _setOptionsByReference() {
-        super._setOptionsByReference();
-
-        extend(this._optionsByReference, {
-            item: true
-        });
-    }
-}
