+++ package.json
@@ -64,7 +64,7 @@
     "cssom": "^0.4.4",
     "del": "^2.2.2",
     "devextreme-cldr-data": "^1.0.2",
-    "devextreme-generator": "^1.0.88",
+    "devextreme-generator": "^1.0.91",
     "devextreme-internal-tools": "stable",
     "enzyme": "^3.11.0",
     "enzyme-adapter-preact-pure": "^2.2.0",
